requirements.txt
openai>=1.23.0
python-dotenv>=1.0.0
Flask>=3.0.0
duckduckgo-search>=5.3.0
flask
requests
gunicorn